+++ package.json
@@ -23,7 +23,6 @@
     "expo-status-bar": "~1.3.0",
     "firebase": "^9.8.4",
     "geofire-common": "^5.2.0",
-    "jest": "^26.6.3",
     "jest-config": "^26.6.3",
     "jest-environment-jsdom": "^27.5.1",
     "jest-expo": "^46.0.1",
@@ -39,36 +38,8 @@
     "@babel/core": "^7.12.9",
     "@testing-library/jest-native": "^4.0.5",
     "@testing-library/react-native": "^11.0.0",
-    "jest": "^28.1.3",
+    "jest": "^26.6.3",
     "react-test-renderer": "^17.0.2"
-  },
-  "jest": {
-    "preset": "jest-expo",
-    "setupFilesAfterEnv": [
-      "@testing-library/jest-native/extend-expect"
-    ],
-    "transformIgnorePatterns": [
-      "node_modules/(?!((jest-)?react-native|@react-native(-community)?)|expo(nent)?|@expo(nent)?/.*|@expo-google-fonts/.*|react-navigation|@react-navigation/.*|@unimodules/.*|unimodules|sentry-expo|native-base|react-native-svg)"
-    ],
-    "moduleFileExtensions": [
-      "ts",
-      "tsx",
-      "js",
-      "jsx",
-      "json",
-      "node"
-    ],
-    "moduleDirectories": [
-      "node_modules"
-    ],
-    "testEnvironment": "jsdom",
-    "testEnvironmentOptions": {},
-    "modulePaths": [
-      "<rootDir>"
-    ],
-    "roots": [
-      "<rootDir>"
-    ]
   },
   "private": true
 }
