+++ src/app/components/molecules/AnimatedInput.js
