+++ src/app/services/initializaiton.js
@@ -1,10 +1,10 @@
 import { initializeApp, getApps, getApp } from 'firebase/app';
-import firebaseConfig from '../firebaseConfig'
 import { getAuth, initializeAuth } from "firebase/auth";
 import { getReactNativePersistence } from 'firebase/auth/react-native';
 import AsyncStorage from '@react-native-async-storage/async-storage';
 import { getFirestore } from 'firebase/firestore';
 
+import firebaseConfig from '../../../firebaseConfig'
 
 let fireApp, fireAuth, fireDB;
 export default initialize = () => {
